+++ lib/omq/cli/peer.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module OMQ
-  module CLI
-    class PeerRunner < ServerRunner
-    end
-  end
-end